+++ cmd/soroban-rpc/internal/integrationtest/get_network_test.go
@@ -1,4 +1,4 @@
-package test
+package integrationtest
 
 import (
 	"context"
@@ -6,11 +6,12 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
+	"github.com/stellar/soroban-rpc/cmd/soroban-rpc/internal/integrationtest/infrastructure"
 	"github.com/stellar/soroban-rpc/cmd/soroban-rpc/internal/methods"
 )
 
 func TestGetNetworkSucceeds(t *testing.T) {
-	test := NewTest(t, nil)
+	test := infrastructure.NewTest(t, nil)
 
 	client := test.GetRPCLient()
 
@@ -19,7 +20,7 @@ func TestGetNetworkSucceeds(t *testing.T) {
 	var result methods.GetNetworkResponse
 	err := client.CallResult(context.Background(), "getNetwork", request, &result)
 	assert.NoError(t, err)
-	assert.Equal(t, friendbotURL, result.FriendbotURL)
-	assert.Equal(t, StandaloneNetworkPassphrase, result.Passphrase)
+	assert.Equal(t, infrastructure.FriendbotURL, result.FriendbotURL)
+	assert.Equal(t, infrastructure.StandaloneNetworkPassphrase, result.Passphrase)
 	assert.GreaterOrEqual(t, result.ProtocolVersion, 20)
 }
