+++ cmd/soroban-rpc/internal/integrationtest/get_version_info_test.go
@@ -1,4 +1,4 @@
-package test
+package integrationtest
 
 import (
 	"context"
@@ -6,7 +6,10 @@ import (
 	"os/exec"
 	"testing"
 
+	"github.com/stretchr/testify/require"
+
 	"github.com/stellar/soroban-rpc/cmd/soroban-rpc/internal/config"
+	"github.com/stellar/soroban-rpc/cmd/soroban-rpc/internal/integrationtest/infrastructure"
 
 	"github.com/stretchr/testify/assert"
 
@@ -14,11 +17,11 @@ import (
 )
 
 func TestGetVersionInfoSucceeds(t *testing.T) {
-	test := NewTest(t, nil)
+	test := infrastructure.NewTest(t, nil)
 
 	version, commitHash, buildTimeStamp := config.Version, config.CommitHash, config.BuildTimestamp
 
-	populateVersionInfo(test)
+	populateVersionInfo(t)
 
 	// reset to previous config values
 	t.Cleanup(func() {
@@ -36,26 +39,23 @@ func TestGetVersionInfoSucceeds(t *testing.T) {
 	assert.Equal(t, config.Version, result.Version)
 	assert.Equal(t, config.BuildTimestamp, result.BuildTimestamp)
 	assert.Equal(t, config.CommitHash, result.CommitHash)
-	assert.Equal(t, test.protocolVersion, result.ProtocolVersion)
+	assert.Equal(t, test.GetProtocolVersion(), result.ProtocolVersion)
 	assert.NotEmpty(t, result.CaptiveCoreVersion)
 
 }
 
 // Runs git commands to fetch version information
-func populateVersionInfo(test *Test) {
+func populateVersionInfo(t *testing.T) {
 
 	execFunction := func(command string, args ...string) string {
 		cmd := exec.Command(command, args...)
-		test.t.Log("Running", cmd.Env, cmd.Args)
+		t.Log("Running", cmd.Env, cmd.Args)
 		out, innerErr := cmd.Output()
 		if exitErr, ok := innerErr.(*exec.ExitError); ok {
 			fmt.Printf("stdout:\n%s\n", string(out))
 			fmt.Printf("stderr:\n%s\n", string(exitErr.Stderr))
 		}
-
-		if innerErr != nil {
-			test.t.Fatalf("Command %s failed: %v", cmd.Env, innerErr)
-		}
+		require.NoError(t, innerErr)
 		return string(out)
 	}
