+++ cmd/soroban-rpc/internal/integrationtest/infrastructure/docker/captive-core-integration-tests.cfg
