+++ cmd/soroban-rpc/internal/integrationtest/infrastructure/docker/core-start.sh
