+++ cmd/soroban-rpc/internal/integrationtest/infrastructure/docker/docker-compose.rpc.yml
