+++ cmd/soroban-rpc/internal/integrationtest/infrastructure/util.go
@@ -1,0 +1,26 @@
+package infrastructure
+
+import (
+	"path/filepath"
+	"runtime"
+
+	"github.com/stellar/go/txnbuild"
+)
+
+//go:noinline
+func GetCurrentDirectory() string {
+	_, currentFilename, _, _ := runtime.Caller(1)
+	return filepath.Dir(currentFilename)
+}
+
+func CreateTransactionParams(account txnbuild.Account, op txnbuild.Operation) txnbuild.TransactionParams {
+	return txnbuild.TransactionParams{
+		SourceAccount:        account,
+		IncrementSequenceNum: true,
+		Operations:           []txnbuild.Operation{op},
+		BaseFee:              txnbuild.MinBaseFee,
+		Preconditions: txnbuild.Preconditions{
+			TimeBounds: txnbuild.NewInfiniteTimeout(),
+		},
+	}
+}
