+++ cmd/soroban-rpc/internal/integrationtest/metrics_test.go
@@ -1,4 +1,4 @@
-package test
+package integrationtest
 
 import (
 	"fmt"
@@ -15,11 +15,14 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/stellar/soroban-rpc/cmd/soroban-rpc/internal/config"
+	"github.com/stellar/soroban-rpc/cmd/soroban-rpc/internal/integrationtest/infrastructure"
 )
 
 func TestMetrics(t *testing.T) {
-	test := NewTest(t, nil)
-	metrics := getMetrics(test)
+	test := infrastructure.NewTest(t, nil)
+	metricsURL, err := url.JoinPath(test.GetAdminURL(), "/metrics")
+	require.NoError(t, err)
+	metrics := getMetrics(t, metricsURL)
 	buildMetric := fmt.Sprintf(
 		"soroban_rpc_build_info{branch=\"%s\",build_timestamp=\"%s\",commit=\"%s\",goversion=\"%s\",version=\"%s\"} 1",
 		config.Branch,
@@ -30,12 +33,13 @@ func TestMetrics(t *testing.T) {
 	)
 	require.Contains(t, metrics, buildMetric)
 
-	logger := test.daemon.Logger()
-	err := errors.Errorf("test-error")
+	daemon := test.GetDaemon()
+	logger := daemon.Logger()
+	err = errors.Errorf("test-error")
 	logger.WithError(err).Error("test error 1")
 	logger.WithError(err).Error("test error 2")
 
-	metricFamilies, err := test.daemon.MetricsRegistry().Gather()
+	metricFamilies, err := daemon.MetricsRegistry().Gather()
 	assert.NoError(t, err)
 	var metric *io_prometheus_client.MetricFamily
 	for _, mf := range metricFamilies {
@@ -49,13 +53,11 @@ func TestMetrics(t *testing.T) {
 	assert.GreaterOrEqual(t, val, 2.0)
 }
 
-func getMetrics(test *Test) string {
-	metricsURL, err := url.JoinPath(test.adminURL(), "/metrics")
-	require.NoError(test.t, err)
-	response, err := http.Get(metricsURL)
-	require.NoError(test.t, err)
+func getMetrics(t *testing.T, url string) string {
+	response, err := http.Get(url)
+	require.NoError(t, err)
 	responseBytes, err := io.ReadAll(response.Body)
-	require.NoError(test.t, err)
-	require.NoError(test.t, response.Body.Close())
+	require.NoError(t, err)
+	require.NoError(t, response.Body.Close())
 	return string(responseBytes)
 }
