--- cmd/soroban-rpc/internal/test/upgrade_test.go
@@ -1,122 +1,0 @@
-package test
-
-import (
-	"context"
-	"testing"
-	"time"
-
-	"github.com/stellar/go/keypair"
-	"github.com/stellar/go/txnbuild"
-	"github.com/stellar/go/xdr"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"github.com/stellar/soroban-rpc/cmd/soroban-rpc/internal/db"
-)
-
-func TestUpgradeFrom20To21(t *testing.T) {
-	if GetCoreMaxSupportedProtocol() != 21 {
-		t.Skip("Only test this for protocol 21")
-	}
-	test := NewTest(t, &TestConfig{
-		ProtocolVersion: 20,
-	})
-
-	client := test.GetRPCLient()
-
-	sourceAccount := keypair.Root(StandaloneNetworkPassphrase)
-	address := sourceAccount.Address()
-	account := txnbuild.NewSimpleAccount(address, 0)
-
-	helloWorldContract := getHelloWorldContract(t)
-
-	params := preflightTransactionParams(t, client, txnbuild.TransactionParams{
-		SourceAccount:        &account,
-		IncrementSequenceNum: true,
-		Operations: []txnbuild.Operation{
-			createInstallContractCodeOperation(account.AccountID, helloWorldContract),
-		},
-		BaseFee: txnbuild.MinBaseFee,
-		Preconditions: txnbuild.Preconditions{
-			TimeBounds: txnbuild.NewInfiniteTimeout(),
-		},
-	})
-
-	tx, err := txnbuild.NewTransaction(params)
-	assert.NoError(t, err)
-	sendSuccessfulTransaction(t, client, sourceAccount, tx)
-
-	// Upgrade to protocol 21 and re-upload the contract, which should cause a caching of the contract
-	// estimations
-	test.UpgradeProtocol(21)
-	// Wait for the ledger to advance, so that the simulation library passes the right protocol number
-	rpcDB := test.daemon.GetDB()
-	initialLedgerSequence, err := db.NewLedgerEntryReader(rpcDB).GetLatestLedgerSequence(context.Background())
-	require.Eventually(t,
-		func() bool {
-			newLedgerSequence, err := db.NewLedgerEntryReader(rpcDB).GetLatestLedgerSequence(context.Background())
-			require.NoError(t, err)
-			return newLedgerSequence > initialLedgerSequence
-		},
-		time.Minute,
-		time.Second,
-	)
-
-	params = preflightTransactionParams(t, client, txnbuild.TransactionParams{
-		SourceAccount:        &account,
-		IncrementSequenceNum: true,
-		Operations: []txnbuild.Operation{
-			createInstallContractCodeOperation(account.AccountID, helloWorldContract),
-		},
-		BaseFee: txnbuild.MinBaseFee,
-		Preconditions: txnbuild.Preconditions{
-			TimeBounds: txnbuild.NewInfiniteTimeout(),
-		},
-	})
-
-	tx, err = txnbuild.NewTransaction(params)
-	assert.NoError(t, err)
-	sendSuccessfulTransaction(t, client, sourceAccount, tx)
-
-	params = preflightTransactionParams(t, client, txnbuild.TransactionParams{
-		SourceAccount:        &account,
-		IncrementSequenceNum: true,
-		Operations: []txnbuild.Operation{
-			createCreateContractOperation(address, helloWorldContract),
-		},
-		BaseFee: txnbuild.MinBaseFee,
-		Preconditions: txnbuild.Preconditions{
-			TimeBounds: txnbuild.NewInfiniteTimeout(),
-		},
-	})
-
-	tx, err = txnbuild.NewTransaction(params)
-	assert.NoError(t, err)
-	sendSuccessfulTransaction(t, client, sourceAccount, tx)
-
-	contractID := getContractID(t, address, testSalt, StandaloneNetworkPassphrase)
-	contractFnParameterSym := xdr.ScSymbol("world")
-	params = preflightTransactionParams(t, client, txnbuild.TransactionParams{
-		SourceAccount:        &account,
-		IncrementSequenceNum: true,
-		Operations: []txnbuild.Operation{
-			createInvokeHostOperation(
-				address,
-				contractID,
-				"hello",
-				xdr.ScVal{
-					Type: xdr.ScValTypeScvSymbol,
-					Sym:  &contractFnParameterSym,
-				},
-			),
-		},
-		BaseFee: txnbuild.MinBaseFee,
-		Preconditions: txnbuild.Preconditions{
-			TimeBounds: txnbuild.NewInfiniteTimeout(),
-		},
-	})
-	tx, err = txnbuild.NewTransaction(params)
-	assert.NoError(t, err)
-
-	sendSuccessfulTransaction(t, client, sourceAccount, tx)
-}
